adding different reference wind speeds in GCL python
